+++ test/integration/authentication-oidc-test.js
@@ -1,4 +1,4 @@
-const Solid = require('../../index')
+const server = require('../../index')
 const path = require('path')
 const fs = require('fs-extra')
 const UserAccount = require('../../lib/account-mgmt/user-account')
@@ -59,7 +59,7 @@ describe('Authentication API (OIDC)', () => {
   let bobPod
 
   before(async () => {
-    alicePod = await Solid.createServer(
+    alicePod = await server.createServer(
       Object.assign({
         root: aliceRootPath,
         serverUri: aliceServerUri,
@@ -68,7 +68,7 @@ describe('Authentication API (OIDC)', () => {
 
     await startServer(alicePod, 7000)
 
-    bobPod = await Solid.createServer(
+    bobPod = await server.createServer(
       Object.assign({
         root: bobRootPath,
         serverUri: bobServerUri,
@@ -159,7 +159,9 @@ describe('Authentication API (OIDC)', () => {
       let response, cookie
 
       before(done => {
-        const aliceAccount = UserAccount.from({ webId: aliceWebId })
+        const aliceAccount = UserAccount.from({
+          webId: aliceWebId, username: 'alice'
+        })
 
         aliceCredentialStore.createUser(aliceAccount, alicePassword)
         alice.post('/login/password')
@@ -168,6 +170,10 @@ describe('Authentication API (OIDC)', () => {
           .send({ password: alicePassword })
           .end((err, res) => {
             response = res
+            const error = err | response.error
+            if (error) {
+              return done(error)
+            }
             cookie = response.headers['set-cookie'][0]
             done(err)
           })
@@ -306,8 +312,6 @@ describe('Authentication API (OIDC)', () => {
           // Submitting select-provider form redirects to Alice's pod's /authorize
           const authorizeUri = res.header.location
           expect(authorizeUri.startsWith(aliceServerUri + '/authorize'))
-
-          console.log('REDIRECTED TO:', authorizeUri)
 
           // Follow the redirect to /authorize
           const authorizePath = authorizeUri.replace(aliceServerUri, '') // (new URL(authorizeUri)).pathname