+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   a `.js` file (to enable comments, etc), and not JSON. Default config file
   is now `config.dev.js` instead of `config.json`.
 - Updated `commander` and `inquirer` packages to latest versions.
+- **BREAKING**: Account recover email is now loaded from `storage.users` db,
+  instead of from the account root `.acl` file.
 
 ## 7.0.0 - 2020-02-15